tests/wpt/web-platform-tests/bluetooth/requestDevice/radio-not-present.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
const test_desc = 'Reject with NotFoundError if there is no BT radio present.';
const expected =
    new DOMException('Bluetooth adapter not available.', 'NotFoundError');

bluetooth_test(
    () => navigator.bluetooth.test.simulateCentral({state: 'absent'})
              .then(
                  () => assert_promise_rejects_with_message(
                      requestDeviceWithTrustedClick(
                          {filters: [{services: ['generic_access']}]}),
                      expected, 'Bluetooth adapter is not present.')),
    test_desc);
</script>
